9/24/22
Added scatter tools, building script, minion AI, Pickups

+++ Team2_Project2/Assets/Scripts/Scriptable Objects/WeaponSO.cs
@@ -18,6 +18,6 @@ public class WeaponSO : ScriptableObject
     public AttackStyle secondaryAttackStyle = AttackStyle.Melee;
     public int secondaryDamage = 10;
     public float secondaryRange = 1;
-    public int secondaryAtkSpeed = 1;
+    public float secondaryAtkSpeed = 1;
     [ConditionalValue("secondaryAttackStyle", Condition2 = "Projectile")] public GameObject secondaryProjectile;
 }
